refactor: Creator 클래스 controller 로 이동
- view 에서 값을 받아와 domain 객체를 생성하므로 controller 로 이동함
